lock: move code to repository package
